+++ Nos/AppController.swift
@@ -33,9 +33,12 @@ class AppController: ObservableObject {
     
     func completeOnboarding() {
         currentState = .loggedIn
+        CurrentUser.shared.subscribe(relays: CurrentUser.shared.onboardingRelays)
+
         router.sideMenuPath = NavigationPath()
         router.closeSideMenu()
         router.selectedTab = .discover
+
         analytics.completedOnboarding()
     }
 }