+++ Nos/Views/NoteCard.swift
@@ -28,6 +28,17 @@ struct NoteCard: View {
     var repliesRequest: FetchRequest<Event>
     var replies: FetchedResults<Event> { repliesRequest.wrappedValue }
     
+    @ObservedObject private var currentUser: CurrentUser = .shared
+    
+    @State private var userTappedShowOutOfNetwork = false
+    
+    var showContents: Bool {
+        showFullMessage ||
+        userTappedShowOutOfNetwork ||
+        currentUser.inNetworkAuthors.contains(note.author!) ||
+        Event.discoverTabUserIdToInfo.keys.contains(note.author?.hexadecimalPublicKey ?? "")
+    }
+    
     var replyAvatarUrls: [URL?] {
         var uniqueAuthors: [Author] = []
         var added = Set<Author?>()
@@ -60,13 +71,25 @@ struct NoteCard: View {
     @EnvironmentObject private var relayService: RelayService
     
     private var showFullMessage: Bool
+    private let showReplyCount: Bool
     
-    init(author: Author, note: Event, style: CardStyle = .compact, showFullMessage: Bool = false) {
+    init(
+        author: Author,
+        note: Event,
+        style: CardStyle = .compact,
+        showFullMessage: Bool = false,
+        showReplyCount: Bool = true
+    ) {
         self.author = author
         self.note = note
         self.style = style
         self.showFullMessage = showFullMessage
-        self.repliesRequest = FetchRequest(fetchRequest: Event.allReplies(to: note), animation: .default)
+        self.showReplyCount = showReplyCount
+        if showReplyCount {
+            self.repliesRequest = FetchRequest(fetchRequest: Event.allReplies(to: note), animation: .default)
+        } else {
+            self.repliesRequest = FetchRequest(fetchRequest: Event.emptyRequest())
+        }
     }
 
     var body: some View {
@@ -74,38 +97,60 @@ struct NoteCard: View {
             switch style {
             case .compact:
                 HStack(alignment: .center) {
-                    Button {
-                        router.currentPath.wrappedValue.append(author)
-                    } label: {
-                        HStack(alignment: .center) {
-                            AvatarView(imageUrl: author.profilePhotoURL, size: 24)
-                            Text(author.safeName)
-                                .lineLimit(1)
-                                .font(.subheadline)
-                                .foregroundColor(Color.secondaryTxt)
-                                .multilineTextAlignment(.leading)
-                                .frame(maxWidth: .infinity, alignment: .leading)
-                            if let elapsedTime = note.createdAt?.elapsedTimeFromNowString() {
-                                Text(elapsedTime)
+                    if showContents {
+                        Button {
+                            router.currentPath.wrappedValue.append(author)
+                        } label: {
+                            HStack(alignment: .center) {
+                                AvatarView(imageUrl: author.profilePhotoURL, size: 24)
+                                Text(author.safeName)
                                     .lineLimit(1)
-                                    .font(.body)
-                                    .foregroundColor(.secondaryTxt)
+                                    .font(.subheadline)
+                                    .foregroundColor(Color.secondaryTxt)
+                                    .multilineTextAlignment(.leading)
+                                    .frame(maxWidth: .infinity, alignment: .leading)
+                                if let elapsedTime = note.createdAt?.elapsedTimeFromNowString() {
+                                    Text(elapsedTime)
+                                        .lineLimit(1)
+                                        .font(.body)
+                                        .foregroundColor(.secondaryTxt)
+                                }
                             }
                         }
+                        NoteOptionsButton(note: note)
+                    } else {
+                        Spacer()
                     }
-                    NoteOptionsButton(note: note)
                 }
                 .padding(10)
                 Divider().overlay(Color.cardDivider).shadow(color: .cardDividerShadow, radius: 0, x: 0, y: 1)
                 Group {
-                    CompactNoteView(note: note, showFullMessage: showFullMessage)
+                    if showContents {
+                        CompactNoteView(note: note, showFullMessage: showFullMessage)
+                    } else {
+                        VStack {
+                            Text("This user is outside your network.")
+                                .font(.body)
+                                .foregroundColor(.secondaryTxt)
+                                .padding(15)
+                            SecondaryActionButton(title: Localized.show) {
+                                withAnimation {
+                                    userTappedShowOutOfNetwork = true
+                                }
+                            }
+                            .padding(.bottom, 15)
+                        }
+                        .frame(maxWidth: .infinity)
+                    }
                     Divider().overlay(Color.cardDivider).shadow(color: .cardDividerShadow, radius: 0, x: 0, y: 1)
                     HStack {
-                        StackedAvatarsView(avatarUrls: replyAvatarUrls, size: 20, border: 0)
-                        if let replies = attributedReplies {
-                            Text(replies)
-                                .font(.subheadline)
-                                .foregroundColor(Color.secondaryTxt)
+                        if showReplyCount {
+                            StackedAvatarsView(avatarUrls: replyAvatarUrls, size: 20, border: 0)
+                            if let replies = attributedReplies {
+                                Text(replies)
+                                    .font(.subheadline)
+                                    .foregroundColor(Color.secondaryTxt)
+                            }
                         }
                         Spacer()
                         Image.buttonReply
@@ -150,31 +195,16 @@ struct NoteCard: View {
             return 20
         }
     }
-
-//    private var replies: [ImageMetadata] {
-//        Array(message.metadata.replies.abouts.compactMap { $0.image }.prefix(2))
-//    }
-
-//    private var attributedReplies: AttributedString? {
-//        let replyCount = message.metadata.replies.count
-//        let localized = replyCount == 1 ? Localized.Reply.one : Localized.Reply.many
-//        let string = localized.text(["count": "**\(replyCount)**"])
-//        do {
-//            var attributed = try AttributedString(markdown: string)
-//            if let range = attributed.range(of: "\(replyCount)") {
-//                attributed[range].foregroundColor = .primaryTxt
-//            }
-//            return attributed
-//        } catch {
-//            return nil
-//        }
-//    }
 }
 
 struct NoteCard_Previews: PreviewProvider {
     
     static var persistenceController = PersistenceController.preview
     static var previewContext = persistenceController.container.viewContext
+    static var router = Router()
+    static var emptyPersistenceController = PersistenceController.empty
+    static var emptyPreviewContext = emptyPersistenceController.container.viewContext
+    static var emptyRelayService = RelayService(persistenceController: emptyPersistenceController)
     
     static var shortNote: Event {
         let note = Event(context: previewContext)
@@ -217,6 +247,9 @@ struct NoteCard_Previews: PreviewProvider {
             }
             .preferredColorScheme(.dark)
         }
+        .environment(\.managedObjectContext, emptyPreviewContext)
+        .environmentObject(emptyRelayService)
+        .environmentObject(router)
         .padding()
         .background(Color.appBg)
     }