+++ Nos/Views/RepliesView.swift
@@ -1,5 +1,5 @@
 //
-//  ThreadView.swift
+//  RepliesView.swift
 //  Nos
 //
 //  Created by Matthew Lorentz on 2/14/23.
@@ -20,25 +20,47 @@ struct RepliesView: View {
     
     @State private var alert: AlertState<Never>?
     
+    @State private var subscriptionIDs = [String]()
+    
     var repliesRequest: FetchRequest<Event>
+    /// All replies
     var replies: FetchedResults<Event> { repliesRequest.wrappedValue }
     
     var directReplies: [Event] {
-        replies.filter {
-            ($0.eventReferences?.lastObject as? EventReference)?.referencedEvent?.identifier == note.identifier
+        replies.filter { (reply: Event) in
+            guard let eventReferences = reply.eventReferences?.array as? [EventReference] else {
+                return false
+            }
+            
+            let containsRootMarker = eventReferences.contains(where: { (eventReference: EventReference) in
+                eventReference.marker == "root"
+            })
+            
+            let referencesNoteAsRoot = eventReferences.contains(where: { (eventReference: EventReference) in
+                eventReference.eventId == note.identifier && eventReference.marker == "root"
+            })
+            
+            let containsReplyMarker = eventReferences.contains(where: { (eventReference: EventReference) in
+                eventReference.marker == "reply"
+            })
+            
+            let referencesNoteAsReply = eventReferences.contains(where: { (eventReference: EventReference) in
+                eventReference.eventId == note.identifier && eventReference.marker == "reply"
+            })
+            
+            // This is sloppy, but I'm writing it anyway in a rush.
+            // TODO: make sure there isn't a #[0] event reference this is referring to
+            let referencesNoteTheDeprecatedWay = eventReferences.last?.eventId == note.identifier
+            
+            return (referencesNoteAsRoot && !containsReplyMarker) ||
+                referencesNoteAsReply ||
+                (!containsRootMarker && !containsReplyMarker && referencesNoteTheDeprecatedWay)
         }
     }
     
     init(note: Event) {
         self.note = note
-        
-        if let rootReference = (note.eventReferences?.array as? [EventReference])?
-            .first(where: { $0.marker == "root" }),
-            let rootId = rootReference.eventId {
-            self.repliesRequest = FetchRequest(fetchRequest: Event.allReplies(toEventWith: rootId))
-        } else {
-            self.repliesRequest = FetchRequest(fetchRequest: Event.allReplies(to: note))
-        }
+        self.repliesRequest = FetchRequest(fetchRequest: Event.allReplies(to: note))
     }
     
     private var keyPair: KeyPair? {
@@ -47,11 +69,24 @@ struct RepliesView: View {
     
     var note: Event
     
+    func subscribeToReplies() {
+        // Close out stale requests
+        if !subscriptionIDs.isEmpty {
+            relayService.sendCloseToAll(subscriptions: subscriptionIDs)
+            subscriptionIDs.removeAll()
+        }
+        
+        let eTags = ([note.identifier] + replies.map { $0.identifier }).compactMap { $0 }
+        let filter = Filter(kinds: [.text], eTags: eTags)
+        let subID = relayService.requestEventsFromAll(filter: filter)
+        subscriptionIDs.append(subID)
+    }
+    
     var body: some View {
         VStack {
             ScrollView(.vertical) {
                 LazyVStack {
-                    NoteButton(note: note, showFullMessage: true)
+                    NoteButton(note: note, showFullMessage: true, allowsPush: false, showReplyCount: false)
                         .padding(.horizontal)
                     ForEach(directReplies.reversed()) { event in
                         ThreadView(root: event, allReplies: replies.reversed())
@@ -60,14 +95,24 @@ struct RepliesView: View {
                 .padding(.bottom)
             }
             .padding(.top, 1)
-            .navigationBarTitle(Localized.threadView.string, displayMode: .inline)
+            .navigationBarTitle(Localized.thread.string, displayMode: .inline)
             .toolbarBackground(.visible, for: .navigationBar)
             .toolbarBackground(Color.cardBgBottom, for: .navigationBar)
+            .onAppear {
+                subscribeToReplies()
+            }
+            .refreshable {
+                subscribeToReplies()
+            }
+            .onDisappear {
+                relayService.sendCloseToAll(subscriptions: subscriptionIDs)
+                subscriptionIDs.removeAll()
+            }
             VStack {
                 Spacer()
                 VStack {
                     HStack(spacing: 10) {
-                        if let author = CurrentUser.author {
+                        if let author = CurrentUser.shared.author {
                             AvatarView(imageUrl: author.profilePhotoURL, size: 35)
                         }
                         ExpandingTextFieldAndSubmitButton( placeholder: "Post a reply", reply: $reply) {
@@ -98,19 +143,14 @@ struct RepliesView: View {
             }
             
             var tags: [[String]] = [["p", note.author!.publicKey!.hex]]
-            if note.eventReferences?.count ?? 0 > 0 {
-                if let referenceArray = note.eventReferences?.array as? [EventReference],
-                    let firstReference = referenceArray.first {
-                    if let rootReference = referenceArray.first(where: { $0.marker == "root" }) {
-                        tags.append(["e", rootReference.eventId ?? "", "", "root"])
-                        tags.append(["e", note.identifier!, "", "reply"])
-                    } else {
-                        tags.append(["e", firstReference.eventId ?? "", "", "reply"])
-                    }
-                }
+            // If `note` is a reply to another root, tag that root
+            if let rootNoteIdentifier = note.rootNote()?.identifier, rootNoteIdentifier != note.identifier {
+                tags.append(["e", rootNoteIdentifier, "", "root"])
+                tags.append(["e", note.identifier!, "", "reply"])
             } else {
                 tags.append(["e", note.identifier!, "", "root"])
             }
+            
             // print("tags: \(tags)")
             let jsonEvent = JSONEvent(
                 id: "",
@@ -147,19 +187,32 @@ struct RepliesView_Previews: PreviewProvider {
         KeyChain.save(key: KeyChain.keychainPrivateKey, data: Data(KeyFixture.alice.privateKeyHex.utf8))
         return persistenceController
     }()
-    
     static var previewContext = persistenceController.container.viewContext
+    static var emptyPersistenceController = PersistenceController.empty
+    static var emptyPreviewContext = emptyPersistenceController.container.viewContext
+    static var emptyRelayService = RelayService(persistenceController: emptyPersistenceController)
+    static var router = Router()
     
     static var shortNote: Event {
         let note = Event(context: previewContext)
+        note.kind = 1
         note.content = "Hello, world!"
+        note.author = user
         return note
     }
     
     static var longNote: Event {
         let note = Event(context: previewContext)
+        note.kind = 1
         note.content = .loremIpsum(5)
+        note.author = user
         return note
+    }
+    
+    static var user: Author {
+        let author = Author(context: previewContext)
+        author.hexadecimalPublicKey = "d0a1ffb8761b974cec4a3be8cbcb2e96a7090dcf465ffeac839aa4ca20c9a59e"
+        return author
     }
     
     static var previews: some View {
@@ -171,6 +224,9 @@ struct RepliesView_Previews: PreviewProvider {
                 RepliesView(note: longNote)
             }
         }
+        .environment(\.managedObjectContext, previewContext)
+        .environmentObject(emptyRelayService)
+        .environmentObject(router)
         .padding()
         .background(Color.cardBackground)
     }