+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Add self-hosted PostHog analytics
 
+## [0.1 (8)] - 2023-03-13Z
+- Fix translucent tab bar in the simulator.
+- Connect to more relays to get user relay list after importing a key during onboarding
+- Fix some bugs in thread views
+- Show placeholder cards for messages outside 2 hops and allow the user to reveal them
+- Support deprecated reply #e tag format
+- Fixed an issue where older contact list and metadata events could overwrite new ones
+- Styled onboarding views
+
 ## [0.1 (7)] - 2023-03-10Z
 - Use only relays added in RelayView for sending and receiving events
 - Add PostHog analytics